requirements.txt
aiohttp
feedparser
trafilatura
beautifulsoup4
requests
google-genai
nest_asyncio
python-dotenv
